Untracked consumption errors in React renderer
This commit adds infrastructure ("read barrier") that makes it possible
to emit errors if any cell is consumed in a rendering context, but it
is not inside a tracking frame.

It also threads through stack information for both the creation of
reactive variables and for consumption of reactive variables. Making
this a dev-time-only cost is extremely urgent and must be done before
0.6.

This commit also implements the read barrier assertion in the React
renderer. The long and short of it is that any reads that happen inside
of a React render function but outside of a useSetup or useReactive
will trigger the error.

Finally, this commit adds TIMELINE.entryPoint, which allows third-party
code to mask abstractions from developer tools and debug messages.

This allows libraries to help make sure that these messages actually
point at user code and not internals. The `@starbeamx/store` package
now makes use of this infrastructure.