[WIP] Improvements to `@starbeam/react` (#78)
This commit updates `@starbeam/react` in a number of ways:

## [superficial] Renames `useReactiveSetup` to `Component`

I don't know if this is the right name, and plan to add back `useReactiveSetup` as an alias for 0.9, but I want a nice name for the combo of `useSetup` and `useReactive`.

I think it's the right default for most people, because it lets you close over and use the stable values creates in `useSetup` without caveats.

## Adds a `props` parameter to `Component`

The main problem with the original `useReactiveSetup` design is that the closure it returns is created once. This is necessary to allow it to close over the stable setup values.

Simple example:

```tsx
function Count() {
  return useReactiveSetup(() => {
    // `count` and `increment` are stable across the component's entire
    // mounted lifecycle.
    const count = Cell(0);
    const increment = () => count.current++;

    // this closure closes over those stable values naturally, with no
    // possible stale closure problem. Great!
    return () => (
      <>
        <p>{count.current}</p>
        <button onClick={increment}>++</button>
      </>
    );
  });
}
```

However, if we wanted `Count` to take props, and we wanted to use them inside of `useReactiveSetup`, we'd have a problem:

```tsx
function Count({ initial = 0 }: { initial?: number }) {
  return useReactiveSetup(() => {
    const count = Cell(0);
    const increment = () => count.current++;

    // this closure closes over the original value of `initial`,
    // and it won't update if a new value of initial is passed
    // in.
    return () => (
      <>
        <p>{count.current + initial}</p>
        <button onClick={increment}>++</button>
      </>
    );
  });
}
```

The same problem would happen if you used `useState` outside of the `useReactiveSetup`. The JSX closure would never see the update.

This problem doesn't exist if you decompose `useReactiveSetup` into `useSetup` and `useReactive`.

```tsx
function Count({ initial = 0 }: { initial?: number }) {
  const { count, increment } = useSetup(() => {
    const count = Cell(0);
    const increment = () => count.current++;

    return { count, increment };
  });

  // this closure is created on every render, so it sees the updated `initial`.
  return useReactive(() => (
    <>
      <p>{count.current + initial}</p>
      <button onClick={increment}>++</button>
    </>
  );
}
```

This works, and it's nice that you can separate `useReactiveSetup` in this way. It makes the whole system flexible and compositional, and makes it easier to ease into Starbeam in existing components.

But when you're creating a greenfield Starbeam component, it's much nicer to be able to close over the stable values, like in the original design.

### The Updated `Component` API Takes Props

The updated `Component` API resolves this issue by taking props that get passed into the JSX closure:

```tsx
function Count({ initial = 0 }: { initial?: number }) {
  return Component({ initial }, () => {
    // `count` and `increment` are stable across the component's entire
    // mounted lifecycle.
    const count = Cell(0);
    const increment = () => count.current++;

    // like in the `useReactiveSetup` design, this closure can use the stable
    // `count` and `increment` values directly. But now, the `initial` value
    // from outside `Component` is passed into the JSX closure, and is always
    // up to date.
    return ({ initial }) => (
      <>
        <p>{count.current + initial}</p>
        <button onClick={increment}>++</button>
      </>
    );
  });
}
```

In this design, React will re-render the JSX when:

- `count.current` changes
- the `initial` prop changes.

You could use stable functions created in the `Component` as event handlers without `useEvent`, and create functions that close over cells and formulas (like `increment`) without worrying about staleness (because cells and the functions that interact with them are all stable).

You can also conditionally create cells and formulas, conditionally register `layout` or `idle` handlers (and run them in any order). This is because the setup code runs only once, so they don't need to run in the same order on subsequent renders.

Finally, you can use hooks like `useState` inside of the JSX closure, because it runs on every render. The only values you absolutely need to pass as `Component` props in order to see updates in the JSX closure are the component's own props.

Also, all of this works with [reusable state](https://github.com/reactwg/react-18/discussions/19), which is enforced by React 18 strict mode. This means that if React unmounts and remounts `Count`, you get a new copy of the state in `setup`, as if you had a brand new instance of the `Count` component.

There's still some more fine-tuning to do on this API, but I think it's close to something that we should recommend for green-field Starbeam components. We'll need good docs, but that's a given.

## `use` and `service` work inside of `Component`

You can use `use` and `service` inside of `Component`:

```tsx
function Stopwatch() {
  return Component(({ use }) => {
    // inside of `Component`, this returns a reactive value
    const now = use(Clock);

    return () => <p>{now.current}</p>
  });
}
```

They work the same way as the hook versions, but return a reactive value (that you can use inside the JSX closure) rather than a value.

One last note: throughout this comment, I've been talking about "JSX closures" because that's what the examples are doing, but you can return anything you like from the inner closure and the entire `Component` call will return the most recent result every time.

This is perhaps why the name `Component` isn't quite right (or perhaps deserves an alias with different ergonomics for the non-JSX case).