Retools useSetup
This commit introduces a way to use useSetup with the setup/compute
pattern.

I intend for this to be the default way to introduce React users to
Starbeam in the future. This will be coupled with documentation about
the importance of the setup/compute pattern to Starbeam and a way to
package up universal reactive objects that use the pattern.