Introduce Custom/CustomBlueprint
The idea is to introduce a packaged up concept that `useSetup` can use
directly. This clarifies the setup/compute model for basic reactive
objects, and makes it possible to package up custom reactive objects
that renderers can use directly.

TL;DR this is the start of clarifying and implementing a pattern for writing
universal reactive code in a similar manner as we currently support
writing universal resources.

My hope is that resources become a souped up version of Custom, and that
the explanation for universality flows neatly from Custom to Resource to
Service/Modifier.

🤞