Third time's the charm!
The third attempt to reimplement resource on top of the new primitives worked
beautifully.

It supports `ResourceList`, but where the previous implementation worked
by implicitly adopting resources across runs, the new implementation of
`ResourceList` manages the lifetimes of its child resources explicitly.

The code is nearly as compact, in part because  the new resource
implementation is more honest about separating entire-resource state
from per-run state.

There's a bit of logic still left over supporting the adoption use-case,
which I will remove soon, but even with this leftover code, the new
resource implementation is still fairly compact.

Reworking resources is the last part of the current PR, and this commit
largely wraps up the design work on that.

The main remaining piece of work is updating the React and Preact
renderers, which should be straight forward.