Continue cleaning up resources
This commit evolves the new resource API to support more signatures.
It also improves the types of the use function and use method to support
these improvements.

This PR also continues support for resource metadata (state that lives
for the entire lifetime of the resource and can be used for cross-run
state).

TL;DR use() takes an `IntoBlueprint` which is either a
`ResourceBlueprint` or a `ResourceConstructor`. In the next commit, it
will also take a function that returns a `ResourceBlueprint`, which will
bring back `use(() => Counter(start))`.

There are two remaining resource features needed for parity with the
system before the reactive primitive reform in this PR:

- Support for `use(() => Blueprint)
- Some way to provide an initial value to a resource, to support
  frameworks (React) that can't run the resource constructor until after
  the first render.

The second of these is tricky, because we want to avoid complicating
all resources just to help this one case. That said, now that we have
metadata that is provided to the constructor, perhaps you could provide
the initial value as a distinguished kind of metadata.
I'll have to think about it.