Move linting to turbo
This change also updates the eslint files to use relative paths for the
tsconfig files. This was previously necessary to support vscode linting,
but this change also updates the vscode settings to use
`eslint.workingDirectories` setting with `{ "mode": "auto" }`, which
allows vscode to automatically detect the correct working directory for
each file.