Updated internal documentation
This commit continues the process of documenting the core Starbeam
concepts for contributor consumption, using the package READMEs to
document the APIs of the individual packages.

This factoring is not designed for application developers, but it's
a perfect place to outline the structure of the concepts from the
perspective of someone trying to understand how the codebase is
structured.