Setup bench workspace, and compare with Glimmer
Use `pnpm start` to launch vite with production mode.
Use `pnpm start:dev` to compare with dev mode.

Re-fix the CJS build for `@starbeam/debug`

After reverting my prior change to use 'require' to load buffer,
which was already sus in the first place, (but the switch to require
was intended to fix the CJS build for @starbeam/debug -- it just
happened to be the wrong fix), we received an error during the build
for `@starbeam/debug` that "top level await is not supported in CJS,
and we should instead target `es`"

_except_, we intentionally want a cjs build.
in Node environments, `Buffer` already naturally exists, so the code
that `await import`'s the `buffer` package isn't needed.

To make this more clear, I updated the `import-meta` plugin in
`@starbeam/build-support` to take some options, and wired up a
`import.meta.env.CJS` condition that was then used in the `setup.ts`
file for `@starbeam/debug`. This _keeps_ the `Buffer`-global-creating
logic for ESM, but it is omitted for CJS -- resulting in a successful
build for both targets.

Address type issues

Adjust ImportMetaEnv, and re-fix the gate around await import. in the build plugin, meta.env guards _must_ be exact, no [] access