Disable react tests in prod environment
React testing library assumes that `act` is available in testing, but
it's not available in prod.

For now, disable the React prod tests to land resource reform.

We'll hopefully re-enable the tests soon, but it will require finding a
strategy for writing generic acceptance tests in React that work both in
development and production mode.